+++ examples/styling_properties.js
@@ -59,6 +59,23 @@ var docDefinition = {
 		'\n\n',
 		{
 			text: 'Text background pattern', background: ['stripe45d', 'gray']
+		},
+		{
+			text: 'Customize word break:',
+			pageBreak: 'before'
+		},
+		{
+			text: 'DefaultLine\n"BreakBehaviour" "ForATextWithVeryVery" "LongLongWords"',
+			fontSize: 30,
+		},
+		{
+			text: '\n\n',
+			fontSize: 30,
+		},
+		{
+			text: 'BreakAll\n"LineBreakBehaviour" "ForATextWithVeryVery" "LongLongWords"',
+			fontSize: 30,
+			wordBreak: 'break-all',
 		}
 	],
 	patterns: {