+++ src/TextBreaker.js
@@ -5,9 +5,10 @@ import StyleContextStack from './StyleContextStack';
 /**
  * @param {string} text
  * @param {boolean} noWrap
+ * @param {boolean} breakAll
  * @returns {Array}
  */
-const splitWords = (text, noWrap) => {
+const splitWords = (text, noWrap, breakAll = false) => {
 	let words = [];
 	if (text === undefined || text === null) {
 		text = '';
@@ -18,6 +19,14 @@ const splitWords = (text, noWrap) => {
 	if (noWrap) {
 		words.push({ text: text });
 		return words;
+	}
+	if (breakAll) {
+		return text.split('').map(c => {
+			if(c.match(/^\n$|^\r$/)) { // new line
+				return { text: '', lineEnd: true };
+			}
+			return { text: c };
+		});
 	}
 
 	if (text.length === 0) {
@@ -111,16 +120,16 @@ class TextBreaker {
 			let item = texts[i];
 			let style = null;
 			let words;
-
+			let breakAll = StyleContextStack.getStyleProperty(item || {}, styleContextStack, 'wordBreak', 'normal') === 'break-all';
 			let noWrap = StyleContextStack.getStyleProperty(item || {}, styleContextStack, 'noWrap', false);
 			if (isObject(item)) {
 				if (item._textRef && item._textRef._textNodeRef.text) {
 					item.text = item._textRef._textNodeRef.text;
 				}
-				words = splitWords(item.text, noWrap);
+				words = splitWords(item.text, noWrap, breakAll);
 				style = StyleContextStack.copyStyle(item);
 			} else {
-				words = splitWords(item, noWrap);
+				words = splitWords(item, noWrap, breakAll);
 			}
 
 			if (lastWord && words.length) {